consolidate code and hide more internal stuff in react decor
